refactor (moved Order table definition)
