+++ server/db/models/index.js
@@ -1,12 +1,11 @@
 const User = require('./user')
 const Product = require('./product')
+const Order = require('./order')
 
-/**
- * If we had any associations to make, this would be a great place to put them!
- * ex. if we had another model called BlogPost, we might say:
- *
- *    BlogPost.belongsTo(User)
- */
+User.belongsToMany(Product, {through: 'Order'})
+Product.belongsToMany(User, {through: 'Order'})
+Order.belongsTo(Product)
+Order.belongsTo(User)
 
 /**
  * We'll export all of our models here, so that any time a module needs a model,
@@ -17,5 +16,6 @@ const Product = require('./product')
 
 module.exports = {
   User,
-  Product
+  Product,
+  Order
 }
